requirements.txt
Django
django-cors-headers
Pillow
python-dotenv
djangorestframework
djangorestframework-simplejwt
django-mail-templated
python-decouple
djoser

# Testing dependencies
pytest
pytest-django
pytest-cov
factory-boy